+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change log
 
+## 1.1.0 (2023-02-12)
+- feat: add Nunjucks template engine for multiple pages
+- refactor: replace static html files with Nunjucks templates
+- chore: update the html-bundler-webpack-plugin to the latest version
+
+git commit -m 'feat: add Nunjucks template engine for multiple pages'
 ## 1.0.0 (2023-01-18)
 First release